Separate versionable logic from purge old.
i.e. put all versionable logic in the VersionableResource concern. Don't
pollute Importable with this.
