+++ app/importers/concerns/importable.rb
@@ -4,7 +4,7 @@ module Importable
   # can be included into any class that will do so.
 
   included do
-    send(:prepend, ParentInstanceMethods)
+    send(:prepend, Prepend)
 
     class << self
       attr_accessor :disabled
@@ -14,34 +14,14 @@ module Importable
     end
   end
 
-  def available_version
-    nil
-  end
-
-  def stored_version
-    model_class.stored_metadata[:version]
-  end
-
-  module ParentInstanceMethods
+  module Prepend
     def import
       Rails.logger.info "#{self.class.name}: import starting."
-
-      if resource_changed?
-        start_time = Time.now if can_purge_old?
-        super
-        update_metadata available_version
-        model_class.purge_old(start_time) if can_purge_old?
-        Rails.logger.info "#{self.class.name}: import finished (resource updated, new data indexed)."
-      else
-        touch_metadata
-        Rails.logger.info "#{self.class.name}: import finished (resource unchanged, no new data indexed)."
-      end
+      start_time = Time.now if can_purge_old?
+      super
+      model_class.purge_old(start_time) if can_purge_old?
+      Rails.logger.info "#{self.class.name}: import finished."
     end
-  end
-
-  def resource_changed?
-    return true unless available_version
-    stored_version != available_version
   end
 
   def extract_fields(parent_node, path_hash)
@@ -132,5 +112,4 @@ module Importable
   end
 
   delegate :can_purge_old?, to: :model_class
-  delegate :stored_metadata, :update_metadata, :touch_metadata, to: :model_class
 end
