+++ app/importers/screening_list/sdn_data.rb
@@ -1,6 +1,7 @@
 module ScreeningList
   class SdnData
     include Importable
+    include VersionableResource
     include ScreeningList::TreasuryListImporter
     include ScreeningList::MakeNameVariants
     self.default_endpoint = 'http://www.treasury.gov/ofac/downloads/sdn.xml'
