+++ app/importers/screening_list/ssi_data.rb
@@ -1,6 +1,7 @@
 module ScreeningList
   class SsiData
     include Importable
+    include VersionableResource
     include ScreeningList::TreasuryListImporter
     include ScreeningList::MakeNameVariants
     self.default_endpoint = 'http://www.treasury.gov/ofac/downloads/consolidated/consolidated.xml'
